requirements/base.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Authentication & Security
PyJWT==2.8.0
argon2-cffi==23.1.0
django-ratelimit==4.1.0

# Caching
redis==5.0.1
django-redis==5.4.0

# Environment Variables
python-dotenv==1.0.1

# Validation
pydantic==2.5.3

# API Documentation
drf-spectacular==0.27.0

# Utilities
pytz==2024.1
python-dateutil==2.8.2

# Celery (for async tasks like emails)
celery==5.3.6
django-celery-beat==2.6.0

# Monitoring & Logging
sentry-sdk==1.40.0